requirements.txt
attrs
click
cmake
Django
django-allauth
django-cleanup
face-recognition
h5py
json5
oauthlib
pillow
psycopg2-binary
pycparser
PyJWT
python3-openid
pytz
requests
requests-oauthlib
six
sqlparse
urllib3
whitenoise
django-import-export
imutils
xlwt
urllib3
python-dotenv
plotly
django-storages
boto3
djangorestframework
channels-redis
channels
graphene-django
daphne
djangorestframework-jwt
dj-rest-auth
django-auto-prefetching
django-celery-results
djangorestframework-simplejwt
django-cors-headers
django-filter
django-celery-beat
django-debug-toolbar 
eventlet
xlrd
openpyxl